requirements.txt
#faiss-cpu==1.12.0
fastapi==0.116.1
#gunicorn
numpy==2.2.6
openai==1.107.2
pinecone==7.3.0
PyMuPDF==1.26.4
python-dotenv==1.1.1
python-multipart==0.0.20
uvicorn==0.35.0